cleaned up vdpkeepalive testcase and small fix for cleanup in vdpassoc/deassoc

+++ testcases/VdpAssoc.py
@@ -162,6 +162,7 @@ class VdpAssoc(object):
     def cleanup(self):                
         print "Cleanup:"
         skip_proj = False
+        skip_nw = False
         
         try:
             new_project = self.controller.getProject(self.new_tenant)
@@ -188,14 +189,16 @@ class VdpAssoc(object):
                                                          self.new_user, self.new_password)
             if not new_network1:
                 print("Network not found during cleanup")
+                skip_nw = True
         except Exception as e:
             print "Error:", e
-            
-        try:
-            self.controller.deleteNetwork(new_network1['id'], self.new_tenant, 
-                                      self.new_user, self.new_password)
-        except Exception as e:
-            print "Error:", e
+        
+        if skip_nw is False:    
+            try:
+                self.controller.deleteNetwork(new_network1['id'], self.new_tenant, 
+                                          self.new_user, self.new_password)
+            except Exception as e:
+                print "Error:", e
         
         try:
             new_user = self.controller.getUser(self.new_user)
